# Added optimization to ranges::copy/copy_n/copy_backward