--- a/Code/Framework/AzCore/Tests/azcoretests_files.cmake
+++ b/Code/Framework/AzCore/Tests/azcoretests_files.cmake
@@ -223,6 +223,7 @@ set(FILES
     AZStd/Optional.cpp
     AZStd/Pair.cpp
     AZStd/Parallel.cpp
+    AZStd/RangesAlgorithmBenchmarks.cpp
     AZStd/RangesAlgorithmTests.cpp
     AZStd/RangesTests.cpp
     AZStd/RangesUtilityTests.cpp