The `ranges::copy` and `ranges::copy_n` functions will use memcpy if the output type is trivially copyable and both iterators are contiguous.

The `ranges::copy_backward` function will use memmove if the output iterator value type is trivially copyable and both iterators are contiguous.

Also added memcpy/memmove to `ranges::move` and `ranges::move_backward` as well.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

## How was this PR tested?

Verified that the existing [ranges copy](https://github.com/o3de/o3de/blob/development/Code/Framework/AzCore/Tests/AZStd/RangesAlgorithmTests.cpp#L367) test succeed.

Added Benchmark Test to verify the optimizations

```
OKAY Library loaded: D:/o3de/o3de/build/windows/bin/profile/AzCore.Tests.dll
OKAY Symbol found: AzRunBenchmarks
09/30/22 20:20:00
Running D:/o3de/o3de/build/windows/bin/profile/AzTestRunner.exe
Run on (64 X 3700 MHz CPU s)
CPU Caches:
  L1 Data 32K (x32)
  L1 Instruction 32K (x32)
  L2 Unified 524K (x32)
  L3 Unified 16777K (x8)
---------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                   Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------------------------------
RangesAlgorithmBenchmarkFixture/BM_RangesCopy_TriviallyCopyableType_DirectCall                           2.34 ns         2.35 ns    298666667
RangesAlgorithmBenchmarkFixture/BM_RangesCopyN_TriviallyCopyableType_DirectCall                          2.09 ns         2.05 ns    320000000
RangesAlgorithmBenchmarkFixture/BM_RangesCopyBackward_TriviallyCopyableType_DirectCall                   2.09 ns         2.09 ns    344615385
RangesAlgorithmBenchmarkFixture/BM_RangesCopy_TriviallyCopyableType_CompileTime_DirectCall               1.31 ns         1.29 ns    448000000
RangesAlgorithmBenchmarkFixture/BM_RangesCopyN_TriviallyCopyableType_CompileTime_DirectCall              1.23 ns         1.23 ns    407272727
RangesAlgorithmBenchmarkFixture/BM_RangesCopyBackward_TriviallyCopyableType_CompileTime_DirectCall       1.27 ns         1.27 ns    640000000
RangesAlgorithmBenchmarkFixture/BM_RangesCopy_TriviallyCopyableType_MemcpyCall                           2.09 ns         2.09 ns    344615385
RangesAlgorithmBenchmarkFixture/BM_RangesCopy_TriviallyCopyableType_MemmoveCall                          2.07 ns         2.09 ns    344615385
RangesAlgorithmBenchmarkFixture/BM_RangesCopy_TriviallyCopyableType_ForLoopImplementation                20.4 ns         20.4 ns     34461538
OKAY AzRunBenchmarks() returned 0
Returning code: 0
```